requirements-dev.txt
maturin==0.14.8
pytest==6.2.5
pytest-runner==5.2
pytest-pycharm==0.7.0
pytest-mock==3.3.1
tldextract==3.4.0
importnb==0.7.0
pandas==1.5.3
scipy==1.10.1
scikit-learn==1.2.2
jupyter
matplotlib
tabulate
tqdm
datasets
manifest-ml
openai